requirements.txt
streamlit>=1.35.0
langgraph>=0.2.0
langchain-core>=0.2.0
openai>=1.30.0
pinecone>=3.2.0
requests>=2.31.0
python-dotenv>=1.0.0
sec-edgar-downloader>=5.0.0
tiktoken>=0.7.0
feedparser>=6.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
psycopg2-binary>=2.9.9
google-cloud-storage>=2.16.0
cloud-sql-python-connector[pg8000]>=1.9.0
pg8000>=1.31.0
